requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
faiss-cpu
sentence-transformers
langchain
langchain-community
llama-cpp-python
tqdm